list5
Please write a program that asks to enter two sparse vectors, and then calculate their scalar product. The sparse vector in its simplest version can be stored in memory as a pair of vectors - a value vector and an order vector with the same number of dimensions. More memory-saving solutions will be scored higher.